Converts Commit Details into component app
