+++ src/webviews/apps/commitDetails/commitDetails.html
@@ -14,6 +14,9 @@
 				src: url('#{root}/dist/glicons.woff2?d3b316716ee1329763a193a20834cd0a') format('woff2');
 			}
 		</style>
+		<script type="application/javascript" nonce="#{cspNonce}">
+			window.webpackResourceBasePath = '#{webroot}/';
+		</script>
 	</head>
 
 	<body
@@ -21,7 +24,10 @@
 		data-placement="#{placement}"
 		data-vscode-context='{ "preventDefaultContextMenuItems": true, "webview": "#{webviewId}", "webviewInstance": "#{webviewInstanceId}" }'
 	>
-		<gl-commit-details-app id="app"></gl-commit-details-app>
-		#{endOfBody}
+		<gl-commit-details-app
+			name="CommitDetailsView"
+			placement="#{placement}"
+			bootstrap="#{state}"
+		></gl-commit-details-app>
 	</body>
 </html>
